Local Storage removed + usertype new enum file
