+++ lib/pages/auth/dummy_home.dart
@@ -7,7 +7,7 @@ import 'package:heartless/main.dart';
 import 'package:heartless/pages/auth/scan_qr_page.dart';
 import 'package:heartless/pages/chat/contacts_page.dart';
 import 'package:heartless/pages/patient_management/patient_profile_page.dart';
-import 'package:heartless/services/storage/local_storage.dart';
+import 'package:heartless/services/enums/user_type.dart';
 import 'package:heartless/shared/models/app_user.dart';
 import 'package:heartless/shared/provider/auth_notifier.dart';
 import 'package:qr_flutter/qr_flutter.dart';
@@ -46,7 +46,6 @@ class _DummyHomeState extends State<DummyHome> {
     // ! logout function
     void logout() async {
       await _authController.logout(authNotifier);
-      await LocalStorage.clearUser();
       if (context.mounted) {
         // ! ensure that the widget is mounted before navigating
         Navigator.pushNamedAndRemoveUntil(
