+++ lib/pages/auth/login_page.dart
@@ -4,7 +4,6 @@ import "package:flutter_svg/svg.dart";
 import "package:heartless/backend/controllers/auth_controller.dart";
 import "package:heartless/backend/services/notifications/notification_services.dart";
 import "package:heartless/main.dart";
-import 'package:heartless/services/storage/local_storage.dart';
 import "package:heartless/services/phone_auth/phone_auth.dart";
 import 'package:heartless/shared/models/app_user.dart';
 import "package:heartless/shared/provider/auth_notifier.dart";
@@ -69,7 +68,6 @@ class _LoginPageState extends State<LoginPage> {
       log(authNotifier.appUser!.toMap().toString());
       bool success = await _authController.login(authNotifier);
       if (success && context.mounted) {
-        await LocalStorage.saveUser(authNotifier);
         NotificationService.getFirebaseMessagingToken(authNotifier);
         goHome();
       }
@@ -99,7 +97,6 @@ class _LoginPageState extends State<LoginPage> {
     void googleSignIn() async {
       bool success = await _authController.googleSignIn(authNotifier);
       if (success && context.mounted) {
-        await LocalStorage.saveUser(authNotifier);
         NotificationService.getFirebaseMessagingToken(authNotifier);
         goHome();
       }
@@ -142,7 +139,7 @@ class _LoginPageState extends State<LoginPage> {
                       height: 20,
                     ),
                     Text(
-                      '${userTypeToString(authNotifier.userType)} Login',
+                      '${authNotifier.userType.name} Login',
                       style: Theme.of(context).textTheme.headlineMedium,
                     ),
                     const SizedBox(
