+++ lib/services/splash/splash_services.dart
@@ -1,17 +1,21 @@
 import 'dart:developer';
 
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:heartless/services/storage/local_storage.dart';
+import 'package:heartless/backend/services/misc/connect_users.dart';
+import 'package:heartless/shared/models/app_user.dart';
 import 'package:heartless/shared/provider/auth_notifier.dart';
 
 class SplashServices {
   final _auth = FirebaseAuth.instance;
 
-  Future<bool> hasLoggedIn(AuthNotifier authNotifier) async {
+  Future<bool> getUserFromFirebase(AuthNotifier authNotifier) async {
     final user = _auth.currentUser;
+    AppUser? appUser;
     if (user != null) {
-      log("User exists. Checking local storage.");
-      if (await LocalStorage.getUser(authNotifier)) {
+      log("User exists");
+      appUser = await ConnectUsers.getUserDetails(user.uid);
+      if (appUser != null) {
+        authNotifier.setAppUser(appUser);
         log("User data fetched");
         return true;
       } else {
