+++ lib/shared/Models/nurse.dart
@@ -1,3 +1,4 @@
+import 'package:heartless/services/enums/user_type.dart';
 import 'package:heartless/shared/models/app_user.dart';
 
 class Nurse extends AppUser {
