+++ lib/shared/provider/auth_notifier.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/foundation.dart';
+import 'package:heartless/services/enums/user_type.dart';
 import 'package:heartless/shared/models/app_user.dart';
 
 class AuthNotifier with ChangeNotifier {
