+++ lib/widgets/auth/user_type_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:heartless/main.dart';
-import 'package:heartless/shared/models/app_user.dart';
+import 'package:heartless/services/enums/user_type.dart';
 import 'package:heartless/shared/provider/auth_notifier.dart';
 import 'package:heartless/widgets/auth/single_user_type.dart';
